Fix panning/scrolling from one side to the other when entering a level

+++ src/include/gr.h
@@ -32,10 +32,10 @@ extern int shm_flags[shm_maxtbls];
 //	Graphics Driver Header
 
 typedef struct {
-	int vpx, vpy;							// Absolute start wrt Screen
-	int vpxl, vpyl;						// Length
-	int vpox, vpoy;						// Origin Offset
-	int vphi, vpback;
+	int16_t vpx, vpy;							// Absolute start wrt Screen
+	int16_t vpxl, vpyl;						// Length
+	int16_t vpox, vpoy;						// Origin Offset
+	int16_t vphi, vpback;
 	} vptype;
 
 #define varincrx -16304
